+++ packages/eslint-config-airbnb-base/rules/errors.js
@@ -37,7 +37,8 @@ module.exports = {
     'no-ex-assign': 2,
 
     // disallow double-negation boolean casts in a boolean context
-    'no-extra-boolean-cast': 0,
+    // http://eslint.org/docs/rules/no-extra-boolean-cast
+    'no-extra-boolean-cast': 2,
 
     // disallow unnecessary parentheses
     // http://eslint.org/docs/rules/no-extra-parens
@@ -79,7 +80,8 @@ module.exports = {
     'no-sparse-arrays': 2,
 
     // Avoid code that looks like two expressions but is actually one
-    'no-unexpected-multiline': 0,
+    // http://eslint.org/docs/rules/no-unexpected-multiline
+    'no-unexpected-multiline': 2,
 
     // disallow unreachable statements after a return, throw, continue, or break statement
     'no-unreachable': 2,