+++ packages/eslint-config-airbnb-base/rules/es6.js
@@ -24,7 +24,7 @@ module.exports = {
     'arrow-spacing': [2, { before: true, after: true }],
 
     // verify super() callings in constructors
-    'constructor-super': 0,
+    'constructor-super': 2,
 
     // enforce the spacing around the * in generator functions
     // http://eslint.org/docs/rules/generator-star-spacing
@@ -60,7 +60,8 @@ module.exports = {
     'no-restricted-imports': 0,
 
     // disallow to use this/super before super() calling in constructors.
-    'no-this-before-super': 0,
+    // http://eslint.org/docs/rules/no-this-before-super
+    'no-this-before-super': 2,
 
     // disallow useless computed property keys
     // http://eslint.org/docs/rules/no-useless-computed-key
@@ -101,6 +102,8 @@ module.exports = {
     }],
 
     // suggest using Reflect methods where applicable
+    // http://eslint.org/docs/rules/prefer-reflect
+    // TODO: enable
     'prefer-reflect': 0,
 
     // use rest parameters instead of arguments
@@ -108,14 +111,16 @@ module.exports = {
     'prefer-rest-params': 2,
 
     // suggest using the spread operator instead of .apply()
-    'prefer-spread': 0,
+    // http://eslint.org/docs/rules/prefer-spread
+    'prefer-spread': 2,
 
     // suggest using template literals instead of string concatenation
     // http://eslint.org/docs/rules/prefer-template
     'prefer-template': 2,
 
     // disallow generator functions that do not have yield
-    'require-yield': 0,
+    // http://eslint.org/docs/rules/require-yield
+    'require-yield': 2,
 
     // enforce spacing between object rest-spread
     // http://eslint.org/docs/rules/rest-spread-spacing
@@ -123,7 +128,11 @@ module.exports = {
 
     // import sorting
     // http://eslint.org/docs/rules/sort-imports
-    'sort-imports': 0,
+    'sort-imports': [0, {
+      ignoreCase: false,
+      ignoreMemberSort: false,
+      memberSyntaxSortOrder: ['none', 'all', 'multiple', 'single'],
+    }],
 
     // enforce usage of spacing in template strings
     // http://eslint.org/docs/rules/template-curly-spacing